requirements.txt
Django==2.0.4
django-filter==1.1.0
django-mathfilters==0.4.0
djangorestframework==3.7.7
djangorestframework-gis==0.12
Markdown==2.6.11
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.3
six==1.11.0
